+++ app/scripts/move-popup/dimMoveItemProperties.directive.js
@@ -20,7 +20,10 @@
       template: [
         '<div>',
         '<div ng-class="vm.classes">',
-        '  <span ng-if="vm.item.locked" class="locked"></span>',
+        '  <span ng-if="vm.item.lockable" class="fa-stack lock-icon" ng-class="{ locking: vm.locking }" ng-click="vm.setLockState(vm.item);">',
+        '    <i class="fa fa-circle fa-stack-2x"></i>',
+        '    <i class="fa fa-stack-1x fa-inverse" ng-class="{ \'fa-lock\': vm.item.locked, \'fa-unlock\': !vm.item.locked, locking: vm.locking }"></i>',
+        '  </span>',
         '  <span><a target="_new" href="http://db.destinytracker.com/inventory/item/{{vm.item.hash}}">{{vm.title}}</a></span>',
         '  <span ng-if="vm.light" ng-bind="vm.light"></span>',
         '  <span ng-if="::vm.item.weaponClassName" ng-bind="::vm.item.weaponClassName"></span>',
@@ -46,7 +49,7 @@
         '         <span ng-if="stat.bar && stat.value > stat.equippedStatsValue && stat.comparable" class="stat-box-inner" style="width: {{ 100 * stat.equippedStatsValue / stat.maximumValue }}%"></span>',
         '         <span ng-if="stat.bar && stat.value > stat.equippedStatsValue && stat.comparable" class="stat-box-inner higher-stats" style="width: {{ 100 * (stat.value - stat.equippedStatsValue) / stat.maximumValue }}%"></span>',
 
-        '         <span ng-if="!stat.bar && stat.comparable" ng-class="{ \'higher-stats\': (stat.value > stat.equippedStatsValue), \'lower-stats\': (stat.value < stat.equippedStatsValue)}">{{ stat.value }}</span>',
+        '         <span ng-if="!stat.bar && (!stat.equippedStatsName || stat.comparable)" ng-class="{ \'higher-stats\': (stat.value > stat.equippedStatsValue), \'lower-stats\': (stat.value < stat.equippedStatsValue)}">{{ stat.value }}</span>',
         '       </span>',
         '         <span class="stat-box-val" ng-class="{ \'higher-stats\': (stat.value > stat.equippedStatsValue && stat.comparable), \'lower-stats\': (stat.value < stat.equippedStatsValue && stat.comparable)}" ng-show="{{ stat.bar }}" class="lower-stats stat-box-val">{{ stat.value }}</span>',
         '    </div>',
@@ -69,9 +72,9 @@
     };
   }
 
-  MoveItemPropertiesCtrl.$inject = ['$sce', 'dimSettingsService', 'ngDialog', '$scope'];
+  MoveItemPropertiesCtrl.$inject = ['$sce', '$q', 'dimStoreService', 'dimItemService', 'dimSettingsService', 'ngDialog', '$scope', '$rootScope'];
 
-  function MoveItemPropertiesCtrl($sce, settings, ngDialog, $scope) {
+  function MoveItemPropertiesCtrl($sce, $q, storeService, itemService, settings, ngDialog, $scope, $rootScope) {
     var vm = this;
 
     vm.hasDetails = (vm.item.stats && vm.item.stats.length) ||
@@ -79,6 +82,38 @@
       vm.item.objectives;
     vm.showDescription = true;// || (vm.item.description.length &&
                               //    (!vm.item.equipment || (vm.item.objectives && vm.item.objectives.length)));
+    vm.locking = false;
+
+    vm.setLockState = function setLockState(item) {
+      if (vm.locking) {
+        return;
+      }
+      var storeId = item.owner;
+      var storePromise;
+
+      if (storeId === 'vault') {
+        storePromise = $q.when(storeService.getStores())
+          .then(function(stores) {
+            return stores[0];
+          });
+      } else {
+        storePromise = storeService.getStore(item.owner);
+      }
+
+      vm.locking = true;
+
+      storePromise
+        .then(function(store) {
+          return itemService.setLockState(item, store, !item.locked)
+            .then(function(lockState) {
+              item.locked = lockState;
+              $rootScope.$broadcast('dim-filter-invalidate');
+            })
+            .finally(function() {
+              vm.locking = false;
+            });
+        });
+    };
 
     vm.classes = {
       'item-name': true,
@@ -119,7 +154,7 @@
         var item = _.find(items, function(item) {
           return item.equipped && item.type === vm.item.type;
         });
-        if (item) {
+        if (item && vm.item.stats) {
           for (var key in Object.getOwnPropertyNames(vm.item.stats)) {
             var itemStats = item.stats && item.stats[key];
             if (itemStats) {